+++ lib/src/models/config_model.g.dart
@@ -6,7 +6,7 @@ part of 'config_model.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_Config _$$_ConfigFromJson(Map<String, dynamic> json) => _$_Config(
+_$ConfigImpl _$$ConfigImplFromJson(Map<String, dynamic> json) => _$ConfigImpl(
       viewsPath: json['views_path'] as String? ?? 'presentation',
       servicesPath: json['services_path'] as String? ?? 'services',
       widgetsPath: json['widgets_path'] as String? ?? 'presentation',
@@ -17,7 +17,8 @@ _$_Config _$$_ConfigFromJson(Map<String, dynamic> json) => _$_Config(
       lineLength: json['line_length'] as int? ?? 80,
     );
 
-Map<String, dynamic> _$$_ConfigToJson(_$_Config instance) => <String, dynamic>{
+Map<String, dynamic> _$$ConfigImplToJson(_$ConfigImpl instance) =>
+    <String, dynamic>{
       'views_path': instance.viewsPath,
       'services_path': instance.servicesPath,
       'widgets_path': instance.widgetsPath,